multi: 4 flips + 11 climbs - polarity, byte params, and three shapes that were not walls
Sixteen functions moved; four went EXACT. Every one had been filed as a
regalloc / EH-cookie / scheduling wall, and in every case the diff was hiding a
real source defect.

Flipped to 100%
  CMulti::RegisterChannel        @0x0baac0  92.71 -> 100  the scan loop seeds and
      bumps its COUNTER first (`for (i=0, p=..; i<4; i++, p++)`), and the name is a
      CString TEMPORARY - retail feeds operator= the ctor's own eax, a named scoped
      local re-leas the slot. Filed as a "/GX EH-state cookie wall".
  CMulti::RecordDropPlayer2      @0x0bb5e0  92.99 -> 100  the third (occupancy) scan
      re-reads GetSize(): the `m_604[slot] = id` store above it aliases the array
      header, so retail reloads m_nSize. Filed as "this/id register pinning".
  CMulti::DetectConnectionConfig @0x0b82e0  85.89 -> 100  two bugs: the channel-name
      latch is the by-value RETURN (`ch0->m_name = GetString5a0();`), and the
      JoinAndRegisterChannel test is failure-first, which swaps the two dtor tails.
  CNetSession::ArmSlot           @0x0c03f0  98.85 -> 100  `parity` is a u8, not an
      i32. The callee's `mov edx,[esp+8]; and edx,0xff` reads the same either way -
      the CALLER settles it: retail ships only dl (`mov dl,..; shl dl,1; push edx`).

Climbed
  WaitForConnect      74.08 -> 99.82   a plain `while (m_58c == 0)`, not
                                       `if (latched) return 1; do {} while (!latched)`
                                       - retail has ONE `mov eax,1` block. Plus a real
                                       deadline variable (`timeGetTime() + 60000`).
  Render              89.04 -> 99.26   signed `% 128` (cdq/xor/sub/and 0x7f/xor/sub),
                                       ArmSlot's u8 param, the m_drainTimer zero-arm,
                                       and `if (fin != 0)` - the `== 0` spelling had
                                       the whole Verify/PumpB tail inverted.
  BroadcastChannelTable 86.74 -> 98.69 the send RESULT is the return value.
  ParseChannelTable   95.34 -> 98.19   indexed row cursor (anchor rows+0 -> rows+2;
                                       retail is rows+1, still open).
  AutoTuneCmdDelay    85.81 -> 97.87   the +1 belongs to the bump; the resend arms are
                                       `<=8 ? 20 : 30`; one `resend` variable stored
                                       once at the bottom (that is what tail-duplicates
                                       the store+call pair and shrink-wraps `push edi`).
  CreateSession       95.36 -> 97.62   default-then-override, not a ternary.
  PumpA               91.40 -> 95.90   five stat-timer decays spell the zero arm first
                                       (MAX 97.98); TickKillCues takes 0, not
                                       g_frameDelta - retail pushes the pinned ebp.
  WaitForOtherPlayers 80.47 -> 94.42   the LAST shared-exit gate must be spelled
                                       `if (count != 0) { body }` falling through to
                                       `ready:`; `goto ready;` makes cl inline the latch.
  OnJoinConfirm       85.77 -> 93.60   the config scratch is char[0x100] (retail's frame
                                       is exactly 0xd8 = 0x100-0x28 larger), the config
                                       string is re-read per call, m_hostIndex latches 3rd.
  NetSetupDlgProc     87.16 -> 89.53   ONE set of buffers at function scope (both
                                       GetDlgItemTextA calls share gameBuf's slot), and
                                       GetDlgItem is its own statement.
  CreateLocalPlayer   95.59 -> 93.84   `pkt.m_flags |= 0x80` is a read-modify-write, not
                                       a store (MAX held; the RMW is unambiguous).

Cleanup
  g_pEndDialog / g_pGetDlgItemTextA / g_pMessageBeep were per-TU extern function-pointer
  decls standing in for real imports. Replaced with ::EndDialog / ::GetDlgItemTextA /
  ::MessageBeep from <Mfc.h> and deleted the three decls - which is also what lets cl
  CSE the repeated import load into a register the way retail does.

Docs: three new patterns (byte-param-caller-ships-only-dl, clamp-decay-zero-arm-first,
gate-falls-through-to-shared-latch) + INDEX lines.

`gruntz build --full` exits 0; MAX high-water 77.88% unchanged; all FATAL gates OK;
cleanliness counters unchanged (reinterpret_casts 437, view debt 0).

+++ include/Net/NetMgr.h
@@ -320,8 +320,11 @@ struct CNetSession {
     i32 SendOne(CNetCmdSlot* s, i32 v);                                       // bfeb0
     void Reconcile();                                                         // c00f0
     i32 Advance();                                                            // c01d0
-    CGruntzCommand* GetSlotPtr(i32 v);    // c0430  id-map fetch (ex "GlyphTable::Get")
-    void ArmSlot(void* node, i32 parity); // c03f0  id-map store (ex "GlyphTable::Set")
+    CGruntzCommand* GetSlotPtr(i32 v); // c0430  id-map fetch (ex "GlyphTable::Get")
+    // `parity` is a BYTE parameter: the caller ships only dl (`mov dl,..; shl dl,1;
+    // push edx`, CMulti::Render) and the callee zero-extends it (`mov edx,[esp+8];
+    // and edx,0xff`) - an i32 parameter forces a caller-side `and edx,0xff`.
+    void ArmSlot(void* node, u8 parity); // c03f0  id-map store (ex "GlyphTable::Set")
     // Checksum @0xc0590: the game-state signature accumulator over the 4x15
     // placed-grunt roster (defined in src/Gruntz/GameChecksum.cpp; ex the
     // "CGameSyncSig" view). Step2437 is a per-frame poke with no bound RVA,
